// src/app/pages/personas/components/personas-table/personas.interface.ts
export interface PersonasInterface {
    claveauth?: string;
    nicknameauth?: string;
    usuarioauth?: string;
    idpersona?: number;
    descripcion: string;
    direccion: string;
    medidasterreno: string;
    medidasconstruccion: string;
    fechainicio: string;
    fechafin: string;
    idtipopersona: number;
    presupuesto: number;
    idrazonsocialcliente: number;
    idrazonsocialcontratista: number;
    idrazonsocialconstructor: number;
    idrazonsocialasociado: number;
    posiciongps: string;
    idestatuspersona: number;
    observaciones: string;
    desctipopersona?: string;
    direccioncliente?: string;
    razonsocialcontratista?: string;
    razonsocialconstructor?: string;
    razonsocialasociado?: string;
    claveestatuspersona?: string;
    tipopersona?: string;
}